ToDo's and organising
